refactor(threeD): rename foxglove-core directory to core
Shorten the shared 3D rendering module path and update imports across
ThreeD, Pose, UrdfDebug, and urdf-preview entrypoint.
