+++ src/features/panels/Pose/PosePanel.tsx
@@ -27,7 +27,7 @@ import * as THREE from 'three';
 import { Line2 } from 'three/examples/jsm/lines/Line2.js';
 import { LineGeometry } from 'three/examples/jsm/lines/LineGeometry.js';
 import { LineMaterial } from 'three/examples/jsm/lines/LineMaterial.js';
-import { TransformTree } from '../ThreeD/foxglove-core/transformTree';
+import { TransformTree } from '../ThreeD/core/transformTree';
 import type { PoseConfig } from './defaults';
 import { buildTrajectoryLineBands, type TrajectoryLineBand } from './trajectory';
