+++ src/features/panels/ThreeD/core/renderables.ts
